for speed limits below 21 kph logic
Enable acceleration for low speed limits. 

If you are traveling in a high speed zone and you are not on the linear deceleration curve or at least 3 m/s close to it then you should accelerate until you have reached within 3 m/s of the constant decleration curve and then follow that till the low speed limit is reached.